\name{in_dir}
\alias{in_dir}
\title{Execute code in temporarily altered environment}
\usage{
in_dir(indir, code)
}
\arguments{
  \item{indir}{temporary directory}

  \item{code}{code to execute}
}
\description{
This function is very similar to the devtools package
}
\author{
Original function form Hadley Wickham
}

