% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{profile}
\alias{profile}
\title{The cozip profile of an archive}
\usage{
profile(source)
}
\arguments{
\item{source}{Local path or URL to the dataset.}
}
\value{
\code{"none"}, \code{"flat"} or \code{"taco"}.
}
\description{
The cozip profile of an archive
}
