% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reader-sql.R
\name{sql}
\alias{sql}
\title{Query a TACO dataset}
\usage{
sql(dataset, query)
}
\arguments{
\item{dataset}{An object returned by \code{\link[=open_dataset]{open_dataset()}}.}

\item{query}{A SQL query over \code{data}, \code{files}, or a metadata level.}
}
\value{
A tibble.
}
\description{
Query a TACO dataset
}
