% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taco-fit.R
\name{taco_fit}
\alias{taco_fit}
\title{Fit TACO Tabular Foundation Model}
\usage{
taco_fit(
  X,
  y,
  row_compression_percentage = 10,
  problem_type = "classification",
  ...
)
}
\arguments{
\item{X}{x training data}

\item{y}{y training data}

\item{row_compression_percentage}{compression percent for rows}

\item{problem_type}{"classification"}

\item{...}{for TACO additional parameters}
}
\value{
a TACO fit
}
\description{
Fit TACO Tabular Foundation Model
}
