% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taco-pred.R
\name{taco_predict}
\alias{taco_predict}
\title{TACO Predictions}
\usage{
taco_predict(fit, data, type = "proba", ...)
}
\arguments{
\item{fit}{a fitted model}

\item{data}{a dataset}

\item{type}{'proba'}

\item{...}{for passing parameters to the predict method}
}
\value{
predictions
}
\description{
TACO Predictions
}
