% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\name{open_dataset}
\alias{open_dataset}
\title{Open a TACO dataset}
\usage{
open_dataset(source)
}
\arguments{
\item{source}{One or more dataset paths or URLs.}
}
\value{
A \code{taco_dataset} with \code{sources}, \code{collection} and \code{contract}.
}
\description{
Open a TACO dataset
}
