% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taco-package.R
\docType{package}
\name{taco-package}
\alias{taco}
\alias{taco-package}
\title{taco: Read TACO Datasets}
\description{
R reader for TACO v3 datasets. Wraps read_taco() in the DuckDB cozip extension, so the contract and the metadata are read the same way in every language.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asterisk-labs/taco}
  \item Report bugs at \url{https://github.com/asterisk-labs/taco/issues}
}

}
\author{
\strong{Maintainer}: Cesar Aybar \email{cesar@asterisk.coop}

}
\keyword{internal}
