% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taco-package.R
\docType{package}
\name{taco-package}
\alias{taco}
\alias{taco-package}
\title{taco: Read TACO Datasets}
\description{
Read TACO v3 datasets through the DuckDB 'cozip' extension. The package exposes dataset contracts and reads ZIP, FOLDER, and TACOCAT containers through one API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asterisk-labs/taco}
  \item Report bugs at \url{https://github.com/asterisk-labs/taco/issues}
}

}
\author{
\strong{Maintainer}: Cesar Aybar \email{cesar@asterisk.coop}

}
\keyword{internal}
