% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taco-package.R
\docType{package}
\name{taco-package}
\alias{taco}
\alias{taco-package}
\title{taco: TACO - AI-Ready Earth Observation Datasets}
\description{
R bindings to libtaco, the reference library for the TACO format.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asterisk-labs/taco}
  \item Report bugs at \url{https://github.com/asterisk-labs/taco/issues}
}

}
\author{
\strong{Maintainer}: Cesar Aybar \email{cesar@asterisk.coop}

}
\keyword{internal}
